pyproject.toml
[project]
name = "wechat-llm-proxy"
version = "0.1.0"  # Release version
description = "WeChat → LLM proxy via iLink Bot API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "loguru>=0.7",
    "qrcode>=7.4",
    "pillow>=10.0",
    "cryptography>=42.0",
    # anthropic-sdk not needed: we use /v1/messages directly via httpx
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
